+++ main.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 )
 
-const version = "0.5.0"
+const version = "0.5.1"
 
 type options struct {
 	dryRun        bool